refactor: DTO request 패키지 구조 변경
