+++ src/main/java/kitchenpos/dto/request/MenuProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.request;
 
 public class MenuProductRequest {
