+++ src/main/java/kitchenpos/dto/request/OrderTableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.request;
 
 public class OrderTableRequest {
