+++ src/test/java/kitchenpos/fixture/OrderFixture.java
@@ -5,8 +5,8 @@ import java.util.ArrayList;
 import java.util.List;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
-import kitchenpos.dto.OrderCreateRequest;
-import kitchenpos.dto.OrderLineItemRequest;
+import kitchenpos.dto.request.OrderCreateRequest;
+import kitchenpos.dto.request.OrderLineItemRequest;
 
 public class OrderFixture {
